+++ R/model-options.R
@@ -4,7 +4,8 @@ assemblerr_options <- function(prm.use_mu_referencing = FALSE,
                        ode.use_general_linear_advans = TRUE,
                        ode.general_nonlinear_advan = "advan13",
                        ode.general_linear_advan = "advan5",
-                       ode.preferred_trans_routines = c("trans2", "trans4")) {
+                       ode.preferred_trans_routines = c("trans2", "trans4"),
+                       default_record.covariance_step = nm_covariance()) {
   return(
     list(
       prm.use_mu_referencing = prm.use_mu_referencing,
@@ -12,7 +13,8 @@ assemblerr_options <- function(prm.use_mu_referencing = FALSE,
       ode.use_general_linear_advans = ode.use_general_linear_advans,
       ode.general_nonlinear_advan = ode.general_nonlinear_advan,
       ode.general_linear_advan = ode.general_linear_advan,
-      ode.preferred_trans_routines = ode.preferred_trans_routines
+      ode.preferred_trans_routines = ode.preferred_trans_routines,
+      default_record.covariance_step = default_record.covariance_step
     )
   )
 }