+++ NEWS.md
@@ -1,3 +1,4 @@
 # assemblerr (development version)
 
+* option to include covariance step record in generated NONMEM model through `render(m, options = assemblerr_options(default_record.covariance_step = nm_covariance()))`
 * Added a `NEWS.md` file to track changes to the package.